requirements.txt
discord.py>=2.4
pillow>=10.4.0
cachetools
pydantic==2.*
typing_extensions>=4.0.0